+++ src/Generator/Snippet.php
@@ -265,9 +265,9 @@ EOD;
      * Allows to query the first record that match the specified conditions
      *
      * @param mixed \$parameters
-     * @return %s|\Phalcon\Mvc\Model\ResultInterface
+     * @return %s|\Phalcon\Mvc\Model\ResultInterface|\Phalcon\Mvc\ModelInterface|null
      */
-    public static function findFirst(\$parameters = null)
+    public static function findFirst(\$parameters = null): ?\Phalcon\Mvc\ModelInterface
     {
         return parent::findFirst(\$parameters);
     }